Add Workflow to Run Coding Metrics (#4)
# Pull Request

## Description

This pull request adds a new GitHub Actions workflow to automate the
collection of coding metrics. The workflow is triggered on pushes to the
`main` branch and runs on a daily schedule. It uses the
`JackPlowman/coding-metrics` action to gather metrics and is configured
with the necessary permissions and environment variables.

**Automation and workflow setup:**

* Added `.github/workflows/coding-metrics.yml` to define a new workflow
for coding metrics, triggered on pushes to `main` and on a daily
schedule.
* Integrated the `JackPlowman/coding-metrics` GitHub Action to collect
coding metrics, with configuration for authentication and debugging.