Add Readme Documentation (#5)
# Pull Request

## Description

This pull request updates the `README.md` to clarify the repository's
purpose and improve documentation. The main change is updating the
repository name and description to accurately reflect its use for
testing the JackPlowman/coding-metrics GitHub Action.

Documentation updates:

* Changed the repository name to "Test Coding Metrics" and added a
description explaining that this repository is used for testing the
coding metrics GitHub Action.
* Updated the table of contents to match the new repository name.