Set Coding Metrics Permissions (#13)
# Pull Request

## Description

This pull request updates the GitHub Actions workflow configuration for
coding metrics. The main change is an adjustment to workflow permissions
to ensure the job has the necessary access to write contents, which is
required for its operation.

Workflow permissions update:

*
[`.github/workflows/coding-metrics.yml`](diffhunk://#diff-6c07998d92ff9275aef3a8510ed4b7a276f2dfc370668025311a4041add7e91cL9-R16):
Changed the global `permissions` to an empty object and moved
permissions configuration into the `coding-metrics` job, granting
`contents: write` to allow the job to write repository contents.