Show Output SVG in README (#15)
# Pull Request

## Description

This pull request updates the documentation in the `README.md` file to
provide clearer information about the output of the coding-metrics
action. The most important change is the addition of a new section
describing and visually demonstrating the output generated by the tool.

Documentation improvements:

* Added a new section titled `coding-metrics Output` to the `README.md`,
including a description and example SVG output image to help users
understand what the action produces.
* Updated the table of contents to reference the new `coding-metrics
Output` section for easier navigation.