Improve docs (#51)
# Pull Request

## Description

This pull request updates the `README.md` to clarify and standardize the
naming and description of the output produced by the coding-metrics
action. The changes improve clarity for users and contributors regarding
what the action generates and where to find the output.

Documentation improvements:

* Renamed the "coding-metrics Output" section and its references to
"JackPlowman/coding-metrics Output" for consistency with the repository
name.
* Updated the description to specify that the SVG file is named
`output.svg` and located in the root of the repository, providing
clearer instructions for users.